initial FilterTags.jsx component
